Style tweaks and updated "Things to Implement" list in Readme.
- Added function to allocate memory of widgets. Still in testing though. Seems
  to do weird things with the Frame.
- Added icon to username and password entry. Added those icons to the icon
  directory.
- Reduced sleep time when running in preview mode.
- Cleaned up css stylesheets.
- Added style sheet for Frame so that color is not hardcoded.
- Need to implement config file that is read at runtime.

+++ include/elmapp.h
@@ -17,12 +17,11 @@
 #ifndef ELM_APP_H
 #define ELM_APP_H
 
-/* Includes */
+/* Includes (added more for users building their own apps) */
 #include "elmdef.h"
 #include "elmio.h"
 #include <stdbool.h>
 #include <stddef.h>
-#include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
@@ -47,8 +46,8 @@ typedef struct ElmApp
 {
     GtkWidget *     (*display)(ElmCallback);
     enum ElmGravity   gravity;
-    int16_t           x;
-    int16_t           y;
+    unsigned int      x;
+    unsigned int      y;
 } ElmApp;
 
 #endif /* ELM_APP_H */
